requirements.txt
mcp
google-genai
anthropic
python-dotenv